stories-assistant: fix double-entered prompt in chat play helper
A submit that lands while the thread renders past the retry window made
the next attempt re-type the prompt; that Enter is dropped during
streaming, stranding a duplicate in the composer. submitPrompt now
re-checks for submission before typing and clears any leftover, and is
shared from testing/ by the Sketch and Uml stories.
